token bug fix, when unloggeg user goes to account page- shows login windov, myAccTabs refactoring
